requirements.txt
asgiref==3.6.0
Django>=4.2,<5.0
gunicorn==20.1.0
whitenoise==6.4.0
requests==2.31.0 
django-allauth==0.54.0
django-widget-tweaks==1.4.12
Pillow==9.5.0
psycopg2-binary==2.9.6
requests-pkcs12==1.15 
debugpy==1.8.0